test: split exim unit and integration tests
